# frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dune Panel API",
    "version": "3.x",
    "description": "Programmatic interface to a Dune panel. Authenticate either by logging in (cookie) or with an API token from Settings → Security → API Token (Bearer). All endpoints under /panel/api/* honour both modes — an API token is a full-admin credential, so treat it like the panel password."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current panel (basePath aware)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token from Settings → Security → API Token. Send as `Authorization: Bearer <token>`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dune",
        "description": "Session cookie set by POST /login. Browser-only."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/panel/api/clients/list/paged": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Filter, sort, and paginate clients on the server. Each item is a slim row (no uuid/password/auth/flow/security/reverse/tgId) so the clients page can ship 25-ish rows in a few KB instead of the full table. The response also includes a summary computed across the full DB row set so dashboard counters stay stable as the user paginates or filters. Page size capped at 200; fetch /get/:email to obtain the full per-client payload for an edit/info modal.",
        "operationId": "get_panel_api_clients_list_paged",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": true,
            "description": "1-indexed page number. Defaults to 1.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": true,
            "description": "Rows per page. Defaults to 25, capped at 200.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": true,
            "description": "Case-insensitive substring match on email / subId / comment.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": true,
            "description": "Status bucket: online | active | deactive | depleted | expiring.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "required": true,
            "description": "Match clients attached to at least one inbound of this protocol (vless, vmess, trojan, shadowsocks, ...).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": true,
            "description": "Sort key: enable | email | inboundIds | traffic | remaining | expiryTime.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": true,
            "description": "ascend or descend.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "items": [
                      {
                        "email": "alice@example.com",
                        "subId": "abcd1234",
                        "enable": true,
                        "totalGB": 53687091200,
                        "expiryTime": 1735689600000,
                        "limitIp": 0,
                        "reset": 0,
                        "inboundIds": [
                          3,
                          5
                        ],
                        "traffic": {
                          "up": 1024,
                          "down": 4096,
                          "enable": true
                        },
                        "createdAt": 1735000000000,
                        "updatedAt": 1735100000000
                      }
                    ],
                    "total": 2000,
                    "filtered": 47,
                    "page": 1,
                    "pageSize": 25,
                    "summary": {
                      "total": 2000,
                      "active": 1850,
                      "online": [
                        "alice@example.com"
                      ],
                      "depleted": [],
                      "expiring": [],
                      "deactive": []
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/get/{email}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Fetch one client by email, including the inbound IDs and external config IDs it is attached to.",
        "operationId": "get_panel_api_clients_get_email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/add": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Create a new client and attach it to one or more inbounds in a single call. Body is JSON. Per-protocol secrets (UUID for VLESS/VMess, password for Trojan/Shadowsocks, auth for Hysteria) are generated server-side when omitted, so callers can send only the universal fields.",
        "operationId": "post_panel_api_clients_add",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "client": {
                  "email": "alice@example.com",
                  "totalGB": 53687091200,
                  "expiryTime": 1735689600000,
                  "tgId": 0,
                  "limitIp": 0,
                  "enable": true
                },
                "inboundIds": [
                  3,
                  5
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "msg": "Client added"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/update/{email}": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Update an existing client by email. Changes propagate to every attached inbound. Body is the JSON client payload — supply the full set of fields you want to keep (the server replaces the row, it does not patch).",
        "operationId": "post_panel_api_clients_update_email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Current client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "email": "alice@example.com",
                "totalGB": 107374182400,
                "expiryTime": 1767225600000,
                "tgId": 123456789,
                "enable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "msg": "Client updated"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/del/{email}": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Delete a client by email. Removes it from every attached inbound and drops its traffic record unless keepTraffic=1 is passed.",
        "operationId": "post_panel_api_clients_del_email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keepTraffic",
            "in": "query",
            "required": true,
            "description": "Pass 1 to retain the xray_client_traffic row after deletion.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "msg": "Client deleted"
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/{email}/attach": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Attach an existing client to one or more additional inbounds. Body is JSON.",
        "operationId": "post_panel_api_clients_email_attach",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "inboundIds": [
                  7,
                  9
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/{email}/detach": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Detach a client from one or more inbounds without deleting the client.",
        "operationId": "post_panel_api_clients_email_detach",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "inboundIds": [
                  5
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
    "/panel/api/clients/{email}/externalLinks": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Replace a client's external links (per-client share links and remote subscription URLs surfaced in their subscription). Sends the full set; the server replaces all rows.",
        "operationId": "post_panel_api_clients_email_externalLinks",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email (unique identifier).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "externalLinks": [
                  {
                    "kind": "link",
                    "value": "vless://uuid@host:443?...#srv",
                    "remark": "DE"
                  },
                  {
                    "kind": "subscription",
                    "value": "https://provider.example/sub/abc",
                    "remark": "Provider"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/bulkAdjust": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Shift expiry and/or traffic quota for many clients in one call. addDays/addBytes may be negative. Clients with unlimited expiry (expiryTime=0) or unlimited traffic (totalGB=0) are skipped for the corresponding field — bulk extend never converts unlimited to limited. Returns the adjusted count and per-email skip reasons.",
        "operationId": "post_panel_api_clients_bulkAdjust",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "emails": [
                  "alice",
                  "bob"
                ],
                "addDays": 30,
                "addBytes": 53687091200
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                },
                "example": {
                  "success": true,
                  "obj": {
                    "adjusted": 2,
                    "skipped": [
                      {
                        "email": "carol",
                        "reason": "unlimited expiry"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/panel/api/clients/resetTraffic/{email}": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Zero out a single client’s up/down counters. Re-enables the client across every attached inbound and pushes the change to Xray (or the remote node) so depleted users can connect again immediately.",
        "operationId": "post_panel_api_clients_resetTraffic_email",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Client email.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "obj": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}